Security improved and minor fixes to auth decorator and local routes

+++ Musify/resources/v1/AccountAccountSongResource.py
@@ -1,11 +1,12 @@
 from flask_restful import Resource
 from Model import database, AccountSong, AccountSongSchema
 from flask import request, Response
-from run import ALLOWED_FILE_SONG_EXTENSIONS, accountSongsDirectory
+from run import ALLOWED_FILE_SONG_EXTENSIONS, ACCOUNT_SONGS_DIRECTORY
 from werkzeug.utils import secure_filename
 from datetime import datetime
 from resources.v1.AuthResource import auth_token
 import sys, os, hashlib
+import audio_metadata
 
 accountsong_schema = AccountSongSchema()
 accountsongs_schema = AccountSongSchema(many=True)
@@ -17,11 +18,15 @@ def allowed_file(filename):
 class AccountAccountSongResource(Resource):
     @auth_token
     def get(self, account, account_id):
+        if account.account_id != account_id:
+            return { "status": "failed", "message": "Unauthorized." }, 401
         accountsongs = AccountSong.query.filter_by(account_id=account_id)
         return { "status": "success", "data": accountsongs_schema.dump(accountsongs).data }, 200
 
     @auth_token
     def post(self, account, account_id):
+        if account.account_id != account_id:
+            return { "status": "failed", "message": "Unauthorized." }, 401
         results = []
         for fileData in request.files:
             file = request.files[fileData]
@@ -30,15 +35,18 @@ class AccountAccountSongResource(Resource):
                 newFileName = hashlib.sha1(
                     (filename + str(datetime.now().timestamp())).encode()
                 ).hexdigest() + "." + file.filename.rsplit('.', 1)[1].lower()
+                file.save(os.path.join(ACCOUNT_SONGS_DIRECTORY, newFileName))
+                audioFileMetadata = audio_metadata.load(ACCOUNT_SONGS_DIRECTORY + "/" + newFileName)
                 newAccountSong = AccountSong(
-                    account_id, filename, 
+                    account_id, 
+                    filename, 
+                    str(round(audioFileMetadata.streaminfo.duration / 60, 2)).replace(".", ":"),
                     newFileName,
                     datetime.today().strftime("%Y-%m-%d")
                 )
                 database.session.add(newAccountSong)
                 database.session.commit()
                 result = accountsong_schema.dump(newAccountSong).data
-                file.save(os.path.join(accountSongsDirectory, newFileName))
                 results.append(result)
         if not results:
             return { "status": "failed", "message": "No selected files." }, 400
@@ -46,10 +54,12 @@ class AccountAccountSongResource(Resource):
 
     @auth_token
     def delete(self, account, account_id, account_song_id):
+        if account.account_id != account_id:
+            return { "status": "failed", "message": "Unauthorized." }, 401
         accountSong = AccountSong.query.filter_by(account_song_id=account_song_id, account_id=account_id).first()
         if not accountSong:
-            return { "status": "failed", "message": "This account song does not exist." }, 401
+            return { "status": "failed", "message": "This account song does not exist." }, 422
         database.session.delete(accountSong)
         database.session.commit()
-        os.remove(accountSongsDirectory + "/" + accountSong.song_location)
+        os.remove(ACCOUNT_SONGS_DIRECTORY + "/" + accountSong.song_location)
         return { "status": "success", "message": "Account song deleted." }, 204
